+++ spec/Pim/Bundle/BaseConnectorBundle/Reader/ORM/BulkProductReaderSpec.php
@@ -25,7 +25,7 @@ class BulkProductReaderSpec extends ObjectBehavior
         MetricConverter $metricConverter,
         StepExecution $stepExecution
     ) {
-        $productManager->getFlexibleRepository()->willReturn($repository);
+        $productManager->getProductRepository()->willReturn($repository);
 
         $this->beConstructedWith($productManager, $channelManager, $completenessManager, $metricConverter);