+++ spec/Pim/Bundle/FilterBundle/Filter/ProductValue/ChoiceFilterSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\FilterBundle\Filter\Flexible;
+namespace spec\Pim\Bundle\FilterBundle\Filter\ProductValue;
 
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
@@ -43,7 +43,7 @@ class ChoiceFilterSpec extends ObjectBehavior
         FilterDatasourceAdapterInterface $datasource,
         $utility
     ) {
-        $utility->applyFlexibleFilter($datasource, 'data_name_key', ['foo', 'bar'], 'IN')->shouldBeCalled();
+        $utility->applyFilterByAttribute($datasource, 'data_name_key', ['foo', 'bar'], 'IN')->shouldBeCalled();
 
         $this->apply(
             $datasource,
@@ -61,7 +61,7 @@ class ChoiceFilterSpec extends ObjectBehavior
     ) {
         $collection->count()->willReturn(2);
         $collection->getValues()->willReturn(['foo', 'bar']);
-        $utility->applyFlexibleFilter($datasource, 'data_name_key', ['foo', 'bar'], 'IN')->shouldBeCalled();
+        $utility->applyFilterByAttribute($datasource, 'data_name_key', ['foo', 'bar'], 'IN')->shouldBeCalled();
 
         $this->apply(
             $datasource,
@@ -76,7 +76,7 @@ class ChoiceFilterSpec extends ObjectBehavior
         FilterDatasourceAdapterInterface $datasource,
         $utility
     ) {
-        $utility->applyFlexibleFilter($datasource, 'data_name_key', ['foo', 'bar'], 'NOT IN')->shouldBeCalled();
+        $utility->applyFilterByAttribute($datasource, 'data_name_key', ['foo', 'bar'], 'NOT IN')->shouldBeCalled();
 
         $this->apply(
             $datasource,
@@ -91,7 +91,7 @@ class ChoiceFilterSpec extends ObjectBehavior
         FilterDatasourceAdapterInterface $datasource,
         $utility
     ) {
-        $utility->applyFlexibleFilter($datasource, 'data_name_key', ['foo', 'bar'], 'IN')->shouldBeCalled();
+        $utility->applyFilterByAttribute($datasource, 'data_name_key', ['foo', 'bar'], 'IN')->shouldBeCalled();
 
         $this->apply(
             $datasource,