+++ src/Pim/Bundle/FilterBundle/Filter/ProductFilterUtility.php
@@ -55,7 +55,7 @@ class ProductFilterUtility extends BaseFilterUtility
      */
     public function getProductRepository()
     {
-        return $this->productManager->getFlexibleRepository();
+        return $this->productManager->getProductRepository();
     }
 
     /**
@@ -72,18 +72,18 @@ class ProductFilterUtility extends BaseFilterUtility
     }
 
     /**
-     * Applies filter to query by flexible attribute
+     * Applies filter to query by attribute
      *
      * @param FilterDatasourceAdapterInterface $ds
      * @param string                           $field
      * @param mixed                            $value
      * @param string                           $operator
      */
-    public function applyFlexibleFilter(FilterDatasourceAdapterInterface $ds, $field, $value, $operator)
+    public function applyFilterByAttribute(FilterDatasourceAdapterInterface $ds, $field, $value, $operator)
     {
 
         $attribute  = $this->getAttribute($field);
-        $repository = $this->productManager->getFlexibleRepository();
+        $repository = $this->productManager->getProductRepository();
         if ($attribute) {
             $repository->applyFilterByAttribute($ds->getQueryBuilder(), $attribute, $value, $operator);
         } else {