+++ src/Pim/Bundle/FilterBundle/Filter/ProductValue/AbstractDateFilter.php
@@ -1,20 +1,20 @@
 <?php
 
-namespace Pim\Bundle\FilterBundle\Filter\Flexible;
+namespace Pim\Bundle\FilterBundle\Filter\ProductValue;
 
-use Oro\Bundle\FilterBundle\Filter\AbstractDateFilter;
+use Oro\Bundle\FilterBundle\Filter\AbstractDateFilter as OroAbstractDateFilter;
 use Oro\Bundle\FilterBundle\Form\Type\Filter\DateRangeFilterType;
 use Oro\Bundle\FilterBundle\Datasource\FilterDatasourceAdapterInterface;
 use Pim\Bundle\FilterBundle\Filter\ProductFilterUtility;
 
 /**
- * Flexible filter
+ * Date filter
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-abstract class AbstractFlexibleDateFilter extends AbstractDateFilter
+abstract class AbstractDateFilter extends OroAbstractDateFilter
 {
     /**
      * {@inheritdoc}
@@ -32,7 +32,7 @@ abstract class AbstractFlexibleDateFilter extends AbstractDateFilter
         $dateEndValue = $data['date_end'];
         $type         = $data['type'];
 
-        $this->applyFlexibleDependingOnType(
+        $this->applyFilterDependingOnType(
             $type,
             $ds,
             $dateStartValue,
@@ -46,21 +46,21 @@ abstract class AbstractFlexibleDateFilter extends AbstractDateFilter
     /**
      * {@inheritdoc}
      */
-    protected function applyFlexibleDependingOnType($type, $ds, $dateStartValue, $dateEndValue, $fieldName)
+    protected function applyFilterDependingOnType($type, $ds, $dateStartValue, $dateEndValue, $fieldName)
     {
         switch ($type) {
             case DateRangeFilterType::TYPE_MORE_THAN:
-                $this->applyFlexibleFilterLessMore($ds, $dateStartValue, $fieldName, false);
+                $this->applyFilterByAttributeLessMore($ds, $dateStartValue, $fieldName, false);
                 break;
             case DateRangeFilterType::TYPE_LESS_THAN:
-                $this->applyFlexibleFilterLessMore($ds, $dateEndValue, $fieldName, true);
+                $this->applyFilterByAttributeLessMore($ds, $dateEndValue, $fieldName, true);
                 break;
             case DateRangeFilterType::TYPE_NOT_BETWEEN:
-                $this->applyFlexibleFilterNotBetween($ds, $dateStartValue, $dateEndValue, $fieldName);
+                $this->applyFilterByAttributeNotBetween($ds, $dateStartValue, $dateEndValue, $fieldName);
                 break;
             default:
             case DateRangeFilterType::TYPE_BETWEEN:
-                $this->applyFlexibleFilterBetween($ds, $dateStartValue, $dateEndValue, $fieldName);
+                $this->applyFilterByAttributeBetween($ds, $dateStartValue, $dateEndValue, $fieldName);
                 break;
         }
     }
@@ -68,10 +68,10 @@ abstract class AbstractFlexibleDateFilter extends AbstractDateFilter
     /**
      * {@inheritdoc}
      */
-    protected function applyFlexibleFilterBetween($ds, $dateStartValue, $dateEndValue, $fieldName)
+    protected function applyFilterByAttributeBetween($ds, $dateStartValue, $dateEndValue, $fieldName)
     {
         if ($dateStartValue && $dateEndValue) {
-            $this->util->applyFlexibleFilter(
+            $this->util->applyFilterByAttribute(
                 $ds,
                 $fieldName,
                 array($dateStartValue, $dateEndValue),
@@ -83,15 +83,15 @@ abstract class AbstractFlexibleDateFilter extends AbstractDateFilter
     /**
      * {@inheritdoc}
      */
-    protected function applyFlexibleFilterLessMore($ds, $dateValue, $fieldName, $isLess)
+    protected function applyFilterByAttributeLessMore($ds, $dateValue, $fieldName, $isLess)
     {
-        $this->util->applyFlexibleFilter($ds, $fieldName, $dateValue, $isLess ? '<' : '>');
+        $this->util->applyFilterByAttribute($ds, $fieldName, $dateValue, $isLess ? '<' : '>');
     }
 
     /**
      * {@inheritdoc}
      */
-    protected function applyFlexibleFilterNotBetween($ds, $dateStartValue, $dateEndValue, $fieldName)
+    protected function applyFilterByAttributeNotBetween($ds, $dateStartValue, $dateEndValue, $fieldName)
     {
         $values    = array();
         $operators = array();
@@ -107,7 +107,7 @@ abstract class AbstractFlexibleDateFilter extends AbstractDateFilter
         }
 
         if ($values && $operators) {
-            $this->util->applyFlexibleFilter($ds, $fieldName, $values, $operators);
+            $this->util->applyFilterByAttribute($ds, $fieldName, $values, $operators);
         }
     }
 }