+++ src/Pim/Bundle/FilterBundle/Filter/ProductValue/DateRangeFilter.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace Pim\Bundle\FilterBundle\Filter\Flexible;
+namespace Pim\Bundle\FilterBundle\Filter\ProductValue;
 
 use Oro\Bundle\FilterBundle\Form\Type\Filter\DateRangeFilterType;
 
 /**
- * Flexible filter
+ * Date range filter
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class DateRangeFilter extends AbstractFlexibleDateFilter
+class DateRangeFilter extends AbstractDateFilter
 {
     /**
      * {@inheritdoc}